LayoutTests/fast/text/international/system-language/navigator-language/navigator-language-zh-Hans.html

<!DOCTYPE html><!-- webkit-test-runner [ language=zh-Hans ] -->
<html>
<head>
<script src="../../../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
shouldBeEqualToString("navigator.language", "zh-cn");
</script>
<script src="../../../../../resources/js-test-post.js"></script>
</body>
</html>
